CheckDriveUpdate: separate files
