+++ CheckDriveUpdate/index.ts
@@ -1,13 +1,11 @@
 import {AzureFunction, Context} from "@azure/functions"
 import {google, driveactivity_v2, people_v1, drive_v3} from 'googleapis';
 import {getGoogleClient} from '../utils/google-client';
-import {ignoredActions, japaneseTranslations, colors} from './config';
-import {stripIndent} from 'common-tags';
-import moment from 'moment-timezone';
 import {flatten} from 'lodash';
 import {slack} from '../utils/slack';
-import {fetchDriveItem, DriveItem} from './drive-api';
-import {cacheCalls} from '../utils/utils';
+import {fetchDriveItem} from './drive-api';
+import {Clients, rootFolderId, getDriveItemId} from './lib';
+import {notifyToSlack} from './notify-to-slack';
 
 // import {promises as fs} from 'fs';
 // import path from 'path';
@@ -17,15 +15,6 @@ const main: AzureFunction = async (context: Context, timer: any, lastDate: {ts: 
 };
 
 export default main;
-
-const rootFolderId = process.env.GOOGLE_ROOT_FOLDER_ID;
-
-interface Clients {
-    slack: typeof slack;
-    drive: drive_v3.Drive;
-    driveActivity: driveactivity_v2.Driveactivity;
-    people: people_v1.People;
-}
 
 const fetchAllDriveActivities = async (
     driveActivity: driveactivity_v2.Driveactivity,
@@ -70,166 +59,6 @@ const fetchAllDriveActivities = async (
     return activities;
 };
     
-const getActionName = (actionDetail: driveactivity_v2.Schema$ActionDetail): string =>
-Object.keys(actionDetail)[0];
-
-const getTargetName = (target: driveactivity_v2.Schema$Target): string => {
-    if (target.driveItem) return target.driveItem.title;
-    else if (target.drive) return target.drive.title;
-    else if (target.fileComment) return target.fileComment.parent.title;
-    // else {
-    //     const _exhaustiveCheck: never = target;
-    //     return _exhaustiveCheck;
-    // }
-    // Target is not union...
-};
-
-// TODO: use batchGet https://developers.google.com/people/api/rest/v1/people/getBatchGet
-const getPersonName = async (client: people_v1.People, resourceName: string): Promise<string> => {
-    const person: people_v1.Schema$Person = (await client.people.get({resourceName, personFields: 'names' })).data;
-    const name = person.names?.[0]?.displayName;
-    return name ?? resourceName; // TODO: a better approach?
-};
-
-const formatDate = (timestamp: string): string =>
-    moment(timestamp).tz('Asia/Tokyo').format('YYYY/MM/DD HH:mm:ss');
-
-const getDate = (activity: driveactivity_v2.Schema$DriveActivity): string => {
-    if (activity.timestamp) {
-        return formatDate(activity.timestamp);
-    } else if (activity.timeRange) {
-        return `${formatDate(activity.timeRange.startTime)} - ${formatDate(activity.timeRange.endTime)}`
-    }
-};
-
-const getDriveItemId = (target: driveactivity_v2.Schema$Target): string => {
-    let itemName: string;
-    if (target.driveItem) itemName = target.driveItem.name;
-    else if (target.drive) itemName = target.drive.name;
-    else if (target.fileComment) itemName = target.fileComment.parent.name;
-    // else {
-    //   const _exhaustiveCheck: never = target;
-    //   itemName = _exhaustiveCheck;
-    // }
-    return itemName.replace(/^items\//, '');
-};
-
-/**
- * cache for getPath
- * folderId => path
- */
-const paths: Map<string, Promise<{path: string | null, valid: boolean}>> = new Map([[rootFolderId, Promise.resolve({path: '/', valid: true})]]);
-const getPath = async (client: drive_v3.Drive, item: DriveItem): Promise<string> => {
-    /**
-     * path: path to item
-     * valid: true if path has rootFolderId in ancestor
-     */
-    const rec = cacheCalls(async (client: drive_v3.Drive, folderId: string): Promise<{path: string | null; valid: boolean}> => {
-        const folder = await fetchDriveItem(client, folderId);
-        if (!folder.content.parents) {
-            // folder is root of drive
-            return { path: null, valid: false };
-        }
-        // I want to use Promise.any...
-        const parentPaths = (
-            (await Promise.all(folder.content.parents.map(async parentId => rec(client, parentId))))
-            .filter(({valid}) => valid).map(({path}) => path)
-        );
-        if (parentPaths.length > 0) {
-            const isFolder = folder.content.mimeType === 'application/vnd.google-apps.folder';
-            // folder has one or more parents that is child of rootFolderId
-            return { path: `${parentPaths[0]}${folder.content.name}${isFolder? '/': ''}`, valid: true };
-        } else {
-            // rootFolderId is not ancestor of folderId...
-            return { path: null, valid: false };
-        }
-    }, (c, id) => id, paths);
-    const path = await rec(client, item.content.id);
-    return path.valid? path.path: item.content.name;
-};
-
-const isIgnored = cacheCalls(async (client: drive_v3.Drive, itemId: string): Promise<boolean> => {
-    const item = await fetchDriveItem(client, itemId);
-    if (item.content.parents.length === 0) {
-        // we don't need to ignore this
-        return false;
-    }
-    const isParentsIgnored = await Promise.all(item.content.parents.map(async parentId => isIgnored(client, parentId)));
-    // 全ての親がignoredならtrue
-    return isParentsIgnored.every((x) => x);
-}, (c, id) => id, new Map([
-    ...(process.env.GOOGLE_DRIVE_IGNORED_IDS.split(',').map(s => [s.trim(), Promise.resolve(true)] as [string, Promise<boolean>])),
-    [rootFolderId, Promise.resolve(false)],
-]));
-
-const getEmoji = (item: DriveItem): string => {
-    switch (item.content.mimeType) {
-        case 'application/vnd.google-apps.folder':
-            return ':file_folder:';
-        default:
-            return ':file:';
-    }
-};
-
-const notifyActivity = async ({slack, drive, people: peopleAPI}: Clients, activity: driveactivity_v2.Schema$DriveActivity, drivelogId: string) => {
-    // not notified in order!
-    // maybe chat.scheduleMessage is useful to imitate notification in order
-    // but I think the inorderness is ignorable if the frequency of execution is high enough
-    const actionName = getActionName(activity.primaryActionDetail);
-    if (ignoredActions.includes(actionName)) {
-        return;
-    }
-    const targets: driveactivity_v2.Schema$Target[] = (
-        (await Promise.all(activity.targets.map(
-            async target => ({target, ignored: await isIgnored(drive, getDriveItemId(target))})
-        )))
-        .filter(({ignored}) => !ignored)
-        .map(({target}) => target)
-    );
-    if (targets.length === 0) {
-        return;
-    }
-    const actorsText = (await Promise.all(
-        activity.actors.map(async actor => `${await getPersonName(peopleAPI, actor.user.knownUser.personName)}  さん`)
-    )).join(', ');
-    const text = stripIndent`
-        ${actorsText}が *${activity.targets.length}* 件のアイテムを *${japaneseTranslations[actionName]}* しました。
-        発生日時: ${getDate(activity)}
-    `;
-    let fileURL: string | undefined = undefined;
-    if (targets.length <= 20) {
-        // attachments
-        const attachments = await Promise.all(targets.map(async target => {
-            const item = await fetchDriveItem(drive, getDriveItemId(target));
-            return {
-            color: colors[actionName],
-            title: `${japaneseTranslations[actionName]}: ${getEmoji(item)} ${await getPath(drive, item)}`,
-            text: '', // TODO: include details
-            title_link: item.content.webViewLink
-            };
-        }));
-        await slack.bot.chat.postMessage({
-            channel: drivelogId,
-            text, 
-            icon_emoji: ':google_drive:',
-            username: 'UpdateNotifier',
-            attachments: attachments,
-        });
-    } else {
-        // post snippet
-        fileURL = ((await slack.bot.files.upload({
-            channels: [drivelogId].join(','),
-            content: (await Promise.all(targets.map(
-                async target => {
-                    const item = await fetchDriveItem(drive, getDriveItemId(target));
-                    return `${japaneseTranslations[actionName]}: ${await getPath(drive, item)} (${item.content.webViewLink})`
-                },
-                ))).join('\n'),
-            initial_comment: text,
-        })) as any).file.permalink;
-    }
-}
-
 const addCommentPermission = async ({drive}: Clients, activity: driveactivity_v2.Schema$DriveActivity, groupEmailAddress) => {
     await Promise.all(activity.actions.filter(({detail}) => detail.create).filter(({target}) => target.driveItem?.driveFile).map(async ({target}) => {
         const item = await fetchDriveItem(drive, getDriveItemId(target));
@@ -288,7 +117,7 @@ const checkUpdate = async (context: Context, since: Date): Promise<Date> => {
     const lastChecked = new Date();
     const activities = (await fetchAllDriveActivities(driveActivity, rootFolderId, since)).map(activity => fillEmptyTarget(context, activity));
     const hooks: ((activity: driveactivity_v2.Schema$DriveActivity) =>  unknown)[] = [
-        async (activity) => await notifyActivity(clients, activity, drivelogId),
+        async (activity) => await notifyToSlack(clients, activity, drivelogId, groupEmailAddress),
         async (activity) => await addCommentPermission(clients, activity, groupEmailAddress),
     ]
     await Promise.all(
